fix: hw1 feedback fix - Upgrade python from 3.11 to 3.12 - Use uv to transform the project from app to a library, exposing api only via modules (modules named with "_" to discourage direct use and import) - Add mypy for static type checking - Add pytest-cov for test coverage and reporting - Add circleCI tests and coverage report - Untracked artifact files - The project is capable of being built with build-tools (maigic uses hatchling) and probably being published to PyPi
