+++ LICENSE
@@ -1,6 +1,16 @@
 MIT License
 
-Copyright (c) 2024 Satyam Chatrola
+Copyright (c) 2024 NYU-OSS
+Contributors:
+
+- Kamen Yotov
+- Satyam Chatrola
+- Frank Zhao
+- Sidhved Warik
+- Yaxin Ke
+- Eli Edme
+- Abdullah Suri
+- Mihir Lovekar
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
