--- SlackChatbot/ClientID_Secret.env
@@ -1,3 +1,0 @@
-CLIENT_ID = '7855184757495.7892528086800'
-CLIENT_SECRET = '918fec90decd47b23d25fe58e02a73e1'
-BOT_TOKEN='xoxb-7855184757495-7855216149575-kPNmwbhC4mlAyVyUT7QXWpkJ'
