+++ src/maigic_nyu/__init__.py
@@ -1,0 +1,1 @@
+"""Make files discoverable."""
