+++ src/maigic_nyu/_demo_math.py
@@ -1,0 +1,11 @@
+"""Dummy file to check ruff, pytest and automatic tests with CircleCI."""
+
+
+def addition(num1: int, num2: int) -> int:
+    """Add 2 numebrs and return their sum."""
+    return num1 + num2
+
+
+def substraction(num1: int, num2: int) -> int:
+    """Substract num2 from num1 i.e. num1 -num2."""
+    return num1 - num2
