+++ src/maigic_nyu/api.py
@@ -1,0 +1,4 @@
+"""Maigic API, exposed to the user."""
+
+from maigic_nyu._demo_math import addition as addition  # noqa: PLC0414
+from maigic_nyu._demo_math import substraction as substraction  # noqa: PLC0414
